Updated openstack/openstack Project: openstack/neutron  ebc30229c4eec9f7e025371a0a6a08b10f955041
Log messages for keepalived-state-change in syslog

We don't maintain a long standing log file for the helper process, so
any log messages that we issue from inside it just vanish after router
cleanup.
This patch enforces syslog logging for the daemon, revealing any errors
and other messages that may be triggered by it.

Closes-Bug: #1635629
Change-Id: I8712f5f3a34fc9f403dfba06e7a64bd9eb408c4a