Updated openstack/openstack Project: openstack-infra/storyboard-webclient  9de526ab6aa36b1c7d73a8685d778ddd031dfae7
Fix search controller reloading when a text parameter is added

This replaces the use of $stateParams with parsing the query string
in the given URL instead, using $location. This way the controller
doesn't reload when a text parameter is added to the URL, which was
causing non-text criteria to be lost upon the addition of a text
criterion.

Change-Id: If1fbf8ce4a06538c97aa430cf5c8e703ac453db0
Task: 3493