Updated openstack/openstack Project: openstack/tempest  5a70bd15f23687f64cf3814ac3f8c05de2e5c65c
Fix no module unit test for Python3.6

This commit fixes a unit test of 'test___init___no_module' for
Python3.6. From Python3.6, the error has been changed to
raise a ModuleNotFoundError (subclass of ImportError) when it cannot
find a module[1]. We should keep tempest working on Python3.6 although
we actually don't have py36 gate job now.

So, The testtools.ExpectedException is too strict in this situation. So
I just change it to use assertRaises. This is a bit loose to check
however, I was thinking introducing checking the semver was a bit
overkill here.

[1] https://docs.python.org/3/whatsnew/3.6.html#other-language-changes

Change-Id: I21dc50f20671780f2704d8dd18ca72e0628c8c68
Closes-Bug: #1664454