Updated openstack/openstack Project: openstack/nova  e117260699122608ca034c271120f40e0ae05c3a
conf: Final cleanups in conf/network

The help text for 'use_neutron' didn't reflect reality. This is why we
don't include information about default in the help text, people :).
This option is moved to the top of the file given its importance. A
helpful little note encouraging people not to waste their time on these
opts is also added.

Change-Id: I8ddd0e1153061f112018ecbf13dfc84c21b3a62b
Implements: blueprint centralize-config-options-pike