Updated openstack/openstack Project: openstack/keystone  fed4b35eab3a03ec609ea78bfc5dac91bf9bd97b
Refactor test_revoke to call check_token directly

In the past, we would create a list called events and then try
add the returned calls to the Revoke method to events which would
in turn be looped over in addition to calling is_revoked and then
list_events to check again. Rather than do three types of repetative
calls we can do one call using check_token to check if TokenNotFound
has been raised or if it successfully passed and returned None. This
way we can use check_token, which calls is_revoked with list_events
as a parameter, to validate all this for us and we don't need to do
the additional work of keeping track of a list (events) as well.
First patch towards cleaning up these unit tests in test_revoke.

Change-Id: I815ed5d5229b6a5c93ac8436d55f9d7a40a7194f