Updated openstack/openstack Project: openstack/keystone  f4958a43c53cbf511590396c8fe70dccbc022821
Remove unnecessary processing when deleting grant.

When deleting a grant with an invalid role ID the
exception RoleNotFound will be thrown. However, before
that exception is thrown a lot of unnecessary work
is being done.

In order to improve that process, this patch checks that
the role exists before continuing to delete a grant.

Change-Id: I4f8a26a71096c7dd456195751562cba26c9853d7
Closes-Bug: 1406776