Updated openstack/openstack Project: openstack-infra/project-config  c77a005c7984a4411c40a9da50c54be0c92ad79b
Install more packages with DIB

Eventually we don't want to be running puppet to create nodepool images.
This patch adds all the packages that were declared as puppet package
resources in openstack_project::template[1] to the infra-package-needs
element so that they will exist on the image with our without puppet
managing them.

[1] http://git.openstack.org/cgit/openstack-infra/system-config/tree/modules/openstack_project/manifests/template.pp#n167

Change-Id: Ie1a0aba57390c9c0b269b4cbb076090ae1de73a9