Updated openstack/openstack Project: openstack/oslo.messaging  02cbb6d34614b44fe1c61445c47376d0d306f2ed
Use Sphinx 1.5 warning-is-error

With pbr 2.0 and Sphinx 1.5, the setting for treat sphinx warnings as
errors is setting warning-is-error in build_sphinx section. Migrate
the setting from the old warnerrors one.

The history document is removed because some of the commit messages used
in ChangeLog were being identified as invalid markup and there doesn't
appear to be anyway to edit there retroactively nor disable warnings on
a specific file.

Change-Id: I79e7ac56d5af1151865686761f3d40a11efbf472