Updated openstack/openstack Project: openstack/nova  fb87f64a25155364d9a511e08e4ee048f5c85282
Set size/status during image create with FakeImageService

This is needed for an upcoming change which introduces
a functional test which shelves a server. Shelving a server
creates a snapshot image and in the real world, glance sets
the size and status attributes on the image when it's created
in glance. Our FakeImageService wasn't doing that, so tests
that are running at the same time with the same fake noauth
credentials are listing images and picking up the shelve
snapshot image which doesn't have size or status set and
that produces a KeyError in the API code.

Change-Id: I70d1505c915e80156aaa4d27a7d47e437b40b6f1