Updated openstack/openstack Project: openstack/nova  fd014e6f2f3680d8ee723a5d41158f54adf2e9e2
api-ref: make it clear that os-cells is for cells v1

With cells v2, we need to be clear about what is cells v1 and
what is cells v2 in our documentation. This makes a note in the
os-cells api-ref about that being for cells v1 and also adds a
line saying the entire API is admin-only by default.

Change-Id: I2a06e2c4921ce7172aab33818054dd37b5ef9a65