Updated openstack/openstack Project: openstack/oslo.log  4811a433f3b88969c71dcd168e79ee3d61599e6d
Systemd native journal support

The systemd journal has a native facility for sending logging to it,
which allows for much richer metadata with log messages. These can be
queried later.

This introduces a new 'use_journal' flag, if used will log natively to
the systemd journal. In addition to the existing log messages getting
send to the journal, it also sends a number of additional variables
(when they exist on the context) like REQUEST_ID, which can be
searched with journalctl.

Basic mock testing is included.

Change-Id: If2591576a96d8ca6e36f4dc6600f5412f40509e2