Updated openstack/openstack Project: openstack/nova  e813890f1f3e47d02ae34d146fa48eed9e4b2981
Remove the extension check for os-networks in servers API

The Nova API doesn't support extension anymore. There is no way
to disable any API in the Nova now. So the check for 'os-networks'
is useless now, also as the 'os-networks' is always enabled now,
the check for neutron is useless also.

Partial-implement-blueprint api-no-more-extensions-pike

Change-Id: I393ca6e91d3bd5b4799386db641366e9e4c3a171