Updated openstack/openstack Project: openstack/nova  cea24a9dc4bb8419bf83ba9168877a6fc9a955d7
Remove the json-schema extension point of server create

To get rid of all of stevedore extension point. The
jsons-chema extension point of server create should be
removed. This patch moves all of extended json-schema into
ServerControoler's class attributes 'schema_func_list'.

But the final goal is that merging all of those extended
json-schema into server's main json-schema. So this is a
middle step for all of those refactor for removing stevedore.

Partial-implement-blueprint api-no-more-extensions-pike

Change-Id: Id8ba6a0383cf236259224ace2df3c3d89f82f27d