Updated openstack/openstack Project: openstack/horizon  87f1291de50f0aff5d3dc6a0387efe344bf41b82
Fix the existence check in DataTable Column

Previously DataTable Column checks whether a field exists or not
by checking a result of dict.get(). This has a side effect that
the check complains a field does not exist even when the field
value is None. This commit change the logic to use hasattr to
handle a field with None.

Change-Id: I6ef5ccf90aefd3c407f91a5f3a633d8f68593339
Closes-Bug: #1675912