Updated openstack/openstack Project: openstack/oslo.config  8a0881f217df9f57c7f1e38371d842a404e55417
sphinxext: Handle unescaped percentage signs

The sphinxext module attempts to allow string substitution for default
values in 'Opt.help'. This feature breaks when percentage signs are used
in option help texts and are not escaped. For example, the following
help text from nova:

    help='Probability that peer will become a seeder (1.0 = 100%)'

Simply abort the attempt to substitute text in these cases.

Change-Id: Ia682cd9c91f07fec2faecb4338b777347f774073
Closes-Bug: #1675737