Updated openstack/openstack Project: openstack-dev/grenade  b38248678018d54e770f88ed05e8ced536aa8c14
Source lib/database when upgrading nova

The "database_connection_url" function comes from
lib/database in devstack. We need to source that before
trying to call the function when upgrading nova.

This wasn't caught because of the "|| true" on the map_cell0
call. That can be removed because this nova change fixed it:

aa7b6ebbb254f00fcb548832941ca9dbd3996d9f

Change-Id: Iaa9340feb86df9a942642537474f0a433d8768b7
Closes-Bug: #1670742