Updated openstack/openstack Project: openstack-dev/grenade  fe78c8fe0048d0909ebc74e8a584e2d6560d2024
Updating devstack.localrc.example for accuracy

The devstack.localrc.example file is used by users trying
to figure out how to setup Grenade to run their upgrade testing
but the file seems to be very outdated since it contains some
references to juno and kilo. Also it considers as if Heat is
on by default and Neutron is not, which is no longer true.
This commit attempts to update the example file so it reflects
more accurately the way devstack works currently, and at
the same time tries to be a little bit more explanatory for
new users trying to discover Grenade.

Change-Id: I698131a34dc662b24e9f053c2cc3e1994c859a08