Updated openstack/openstack Project: openstack/tempest  d649055e46b866d290c9a6ef55a3134217caefd6
Skip arbitrary container tests for Ceph

The Ceph backup driver does not support an arbitrary container name
without there being a corresponding pool already configured on the
cluster.  Unlike swift, Ceph cannot create a pool to match a container
name on the fly, and this test is not supported when configured with a
ceph-backed cinder backup service.

This patch fixes commit f62703b5a4ec420399373480531c5d0428454350 which
causes the ceph job to fail.

Change-Id: Ic8d60d58982bb44bfb233aeccc60c35e8eb0aaf3