Updated openstack/openstack Project: openstack/trove  1be92358d96cbb362c0a2300fd77f53156f170fc
add debug information to scenario test stdout

This is to help with debugging gate. The REPORT logging coming from the
test controller and that goes to stdout is not timestamped. This means
that it is hard to correlate it with actual failures that produced it.

Add a timestamp to the [REPORT] lines like so:

2017-04-06 22:26:03.475513 | 2017-04-06 22:26:03.475 | 2017-04-06 22:23:09.931979 [REPORT] Removing replicated data set: DataType.small

Change-Id: I0bad63ec6607978179fd51bd8083723c5115a371