Updated openstack/openstack Project: openstack-infra/project-config  764f11e59c0093f0a4fd83245304df96afc2ca85
Move nodepool mirror vars into seperate filename

Provide a simple way for projects to source nodepool mirror
variables. We construct the new file in such a way that it doesn't
depend on nodepool, but existing nodepool scripts use it just as
downstream consumers would so that we are sure that everything lines up
without double accounting.

Change-Id: I2642659406e72090cffe5163783290e22d6ef3c2
Signed-off-by: Paul Belanger <pabelanger@redhat.com>