Updated openstack/openstack Project: openstack/python-barbicanclient  876d0c854a2f22271caa1cc0b7b3a226a4590079
Enable F and H flake8 rules

Flake8 was configured to ignore F and H rules, which are the PyFlakes
rules and OpenStack hacking rules.

H202 (tests for the too-general Exception used with assertRaises) is
still ignored because some code actually throws a generic Exception
and so the unit test must check that Exception is raised.

Change-Id: I4c4b67631547cd2db784936287355968a6d3b6c7