Updated openstack/openstack Project: openstack-infra/system-config  d9fda5b81f6c8d64fda2ca2c08246492e800292f
Add firehose schema docs

This commit starts a new page for documenting the firehose schema. Right
now it only includes the schema for gerrit events, but it will be
expanded in future patches to include the other services reporting.
Hopefully this will serve as a better guide for people to actually being
able to consume events from firehose.

Change-Id: I2157f702c87f32055ba2fad842a05e31539bc857