Updated openstack/openstack Project: openstack/manila  eaa22371cb7c571d889a488c45e9e81d2c54fc22
Fix dedup/compression description in doc

In the capabilities section of capabilities_and_extra_specs.rst,
the description of dedupe and compression all say that drivers
cannot report to the scheduler both support and non-support,
like [True, False].

But since Mitaka, this kind of specifying is allowed.
So change the doc contents to fit the actual logic.

Change-Id: Ie871abc49bd1bc58566111f85d89228046865b5c
Closes-Bug: #1621707