Updated openstack/openstack Project: openstack-infra/system-config  ff955ff6c9ed6a70486788b743a20fa9ab22b286
Add firehose schema doc for launchpad message

This commit adds schema docs for launchpad messages generated by lpmqtt.
Due to inconsistencies and the lack of a defined schema from launchpad
this doc reads a little non-committal, but that's because nothing is
a guaranteed coming from launchpad and consuming the messages will
require flexibility to accommodate that.

Change-Id: I9b4b1bc27f9bdf81a5fdc53e0969be46a3ded005