Updated openstack/openstack Project: openstack/tempest  07acf26015e5881b76368d90c245b8a16dd8b47f
Move InvalidServiceTag

We are going to make test.py stable for plugins, so it should
only depend on tempest.lib and config.

InvalidServiceTag is in the exceptions module.
It cannot be moved in tempest.lib.exceptions since it has no meaning
in there, so moving it to test.py directly.

Change-Id: I969fb45a44ce08c176d64dfe1c58d75215eacdf5