Updated openstack/openstack Project: openstack/oslo.log  12e3d7944bab27e6e99a9714a22f52cb5e3f4161
Always create OSSysLogHandler

Previously we didn't define OSSysLogHandler at all if it was on a
platform without syslog, because we used syslog at a class level to
build the log priority map.

That now uses a static mapping, which means that OSSysLogHandler can
always be defined, and use the same RuntimeError on init that
OSJournalHandler does.

Change-Id: Id61af05ff10e92e73ec8745be5f1828e6b2effda