Updated openstack/openstack Project: openstack-infra/project-config  5ff7f1b9e336f4c8a4b1a71d8c299bac1317619a
Don't error about default project group for SB

While having a groups entry with the same name as the repository
makes no sense for Launchpad integration, it's entirely reasonable
for projects using StoryBoard to put a project in a group which has
the same name. Stop erroring about a "default groups entry" when
use-storyboard is true.

Change-Id: I95e3a9e830361563776997f2b90d0a48ea66be01