Updated openstack/openstack Project: openstack/openstack-manuals  c8dc4e276e998acb2a2810e456a610ceacb66bac
Fix nova placement API guide for OBS

The nova package in SUSE requires the user to explicitly enable the
Apache vhost for the nova placement API, it is not a part of the
nova-api systemd service nor does it default to enabled. This patch adds
the necessary instructions to ensure the placement API is running on a
SUSE host.

backport: ocata
Change-Id: I104285439c9d46ded4dac0a773648a314741a3e3