Updated openstack/openstack Project: openstack/tempest  c85642f6e2c355669c42572c2b08706ee9f98bbf
Fix heading levels in write_tests doc

This commit fixes the top level header for the write_tests doc.
Previously, the doc title was set as the same as the section. This
caused things like table of contents to confuse the hierarchy of the
sections in the docs. This commit fixes that by setting the doc title
to be a higher lever in the hierarchy.

Change-Id: I2ecb110c766a71e1dc2d9e663ca38ece76dc327d