Updated openstack/openstack Project: openstack-infra/shade  0895729f058370d664dda53fdb11ded19957868e
Don't run extra server info on every server in list

If someone is just doing a get, there is no reason to fill in the extra
information for EVERY server in the list. This makes ansible modules
especially slow.

Instead, if get_server is called, prevent list_servers from doing the
expansion and instead do it late just on the one requested.

Fixes: ansible/ansbile#23354

Change-Id: I42773b292b11f1faa26cf9193b2d0435d9455dd0