Updated openstack/openstack Project: openstack-infra/system-config  88629e814bcf64c7429344bb1aadb133cd33eaa5
Move unbound to openstack_project::server

Once nodepool is installing and fully configuring unbound, we don't need
to manage it in the common openstack_project::template class. This patch
moves them over to openstack_project::server so that it is obvious that
puppet doesn't need to manage them for both single-use workers and
long-lived servers. Eventually the openstack_project::template class
should be empty and we can remove it.

Depends-On: I3905be12acd85581a608d87ba5159cc883343a37
Change-Id: Ie1daae8471d424f373b47262e9d0b21c15affce9