Updated openstack/openstack Project: openstack/keystone-specs  379b9ae3570a7ac6759395b3d9b02d25247b4a5e
Unified limits specification

This is a keystone spec for the conceptual plan of making keystone the
authoritative source for limits in the OpenStack ecosystem. It lays
out the high level approach for doing so, as well as concrete steps to
move this work forward.

This is being proposed to get agreement on direction before moving
forward on the more concrete interfaces. While proposed to the
keystone-specs repository, as the change is mostly going to impact
keystone in early phases, the intent is that a wide range of OpenStack
stakeholders agree on this before moving forward.

Change-Id: I2f338e4c3adceac777cf1e9817f8ec7b7b860abf