Updated openstack/openstack Project: openstack/nova  b87bd4b416567cb71175400604c86b96261358b9
Teach HostAPI about cells

This makes the HostAPI module know enough about cells to support
the os-hypervisors, os-services, and several other API modules that
deal with hosts and services.

Note that this introduces a conflict where duplicate-id resources could
be returned from the API because they are in different DBs with different
key namespaces. For now, we return an error in the delete case where an
ambiguous result could cause us to delete the wrong service. This is
a fundamental problem across several of our APIs and is being worked as
a separate improvement.

Related to blueprint cells-aware-api

Change-Id: If1e03c9343b8cc9c34bd51c2b4d25acdb21131ff