Updated openstack/openstack Project: openstack/diskimage-builder  ffd4820d59dbd1a4000ff3536acb065326714932
Install pip with python interpreter

Currently we install pip/virtualenv with "/usr/local/bin/dib-python".
This means that every time you create a virtualenv, the python
interpreter inside it is called "dib-python" which is confusing.

Add an env var DIB_PYTHON that points directly the to interpreter
available during build, for use when running scripts.

Change-Id: I88ad3c9eb958d58db4631d9b27bc2c592f970345