Updated openstack/openstack Project: openstack/diskimage-builder  7a962e9d1cb4c680dcbc7c73c7080c0bd9e90a39
Skip python3-virtualenv on <= trusty

If DIB_PYTHON_VERSION is < 3 on the !redhat path, that means we're on
an older platform that may not have python3-virtualenv packages.  Skip
install.

Ensure the order of operations happens by forcing the installs

Also add a note about limited platform support (patches welcome :)

Change-Id: I18412767f0ebf946d557a0a126285369e96af159