Updated openstack/openstack Project: openstack-infra/os-loganalyze  bbef755ed80d9e7a5812b01f5e4fd6b70bcb0605
Fix systemd line parsing

There was a bug in systemd parsing with a missing space so that it was
never matching. That's fixed.

In addition, the systemd and syslog line parsing are close enough to
each other, that in our previous patterns they would match each
other's lines. The important distinction is that systemd includes
[pid] after the host which syslog doesn't.

We also have to realize that status might not be in a systemd line,
and that's ok, but still match it, given that otherwise we run the
risk of falling into the syslog match.

Change-Id: Id91c167118894c86b9b0d875a1bb36bdbe239ebd