Updated openstack/openstack Project: openstack/nova  db6382ef043ad475306109420f68e567917cfc1d
Make server_groups determine deleted-ness from InstanceMappingList

The server_groups API currently queries whole instance objects to
determine which of the group members are not deleted. This hits the
cell database and also grabs a ton of data. This patch makes it use
the newly-minted InstanceMappingList to draw the same conclusion,
which is both more effcient and does not require teaching it about
multiple cells.

Related to blueprint cells-aware-api

Change-Id: If571f9e7c9d0ef6265a249ff808d30a24ab462af