Updated openstack/openstack Project: openstack-infra/system-config  3ff96003a86e182297fa94a1b399ad2dbb32297a
Move automatic_upgrades management to o_p::server

single_use_slave turns this off so there is no reason to keep it in the
openstack_project::template class.

This patch also removes the automatic_upgrades parameter from the
single_use_slave class, which is safe because project-config does not
use it.

Change-Id: If4d425cb581f4c5f57fbcdd7eee0622e829cb7ec