Updated openstack/openstack Project: openstack-infra/project-config  c156ac3ef17d382a652cd69f58e8b4c30879670e
Demote Senlin py35 job to non-voting

The initial enabling wasn't as smooth as expected. Seems Devstack is
setting swift as default store backend and swift is not py3 ready.
This problem didn't surface when using py27.

This patch temporarily demotes py35 job to non-voting so that other
patches won't get blocked when we cut p-1 release.

Change-Id: I0d9efb7c2cc023085037f4d1f75b2613a0f1906a