Updated openstack/openstack Project: openstack/nova  8ffacd0efe26b23aba054d7d694802cd0c2cf904
Image meta min_disk should be int in fake_request_spec

This might be blowing up some concurrently running tests
by raising a TypeError comparing an int to a str in py35
jobs. It's hard to reproduce, but we know min_disk should
be an int so let's just change it.

Change-Id: I153544d2aeefbd2d12543481e3bda75ed0316798
Closes-Bug: #1680667