Updated openstack/openstack Project: openstack/neutron-vpnaas  058469e1b99b647537a5228c6a384d93df5484df
VPNaaS integration with services flavor framework

* adds new attribute 'flavor_id' for vpnservice
* adds support for multiple drivers for VPNaaS

This solution uses a flavor of service type VPN associated with a flavor
profile containing a driver. During vpnservice creation, if a flavor_id
is passed, it is used to find the provider for the driver which would
handle the newly created vpnservice.  The flavor_id and the
provider-vpnservice association are pesisted in the DB.

ApiImpact and DocImpact tags are added as the new optional flavor_id
parameter should be described, as well as the support of multiple VPNaaS
drivers.

The original work and discussion about integrating VPNaaS and service
type framework can be found in the following change:
I9e195dfaee21b1cf204cb9b9fc773bc9e5af5936

ApiImpact
DocImpact
Closes-Bug: #1672920

Signed-off-by: Hunt Xu <mhuntxu@gmail.com>

Change-Id: I0095e160481f1d4572e38ad1d3bbc8e183039b84